+++ tuto/Elder.ctxt
@@ -5,5 +5,7 @@ comment1.params=
 comment1.target=void\ act()
 comment1.text=\r\n\ Act\ -\ do\ whatever\ the\ Elder\ wants\ to\ do.\ This\ method\ is\ called\ whenever\r\n\ the\ 'Act'\ or\ 'Run'\ button\ gets\ pressed\ in\ the\ environment.\r\n
 comment2.params=
-comment2.target=void\ createDialog()
-numComments=3
+comment2.target=void\ addHiddenSprite()
+comment3.params=
+comment3.target=void\ elderDialog()
+numComments=4
