+++ tuto/Elder.java
@@ -1,4 +1,5 @@
-import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
+import greenfoot.*;
+import java.util.List;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
 
 /**
  * Write a description of class Elder here.
@@ -8,19 +9,58 @@ import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
  */
 public class Elder extends Actor
 {
+    private HiddenSprite hs;
+    int counter = 10;
+    boolean isActive = false;
+    private TextPanel helloText;
     /**
      * Act - do whatever the Elder wants to do. This method is called whenever
      * the 'Act' or 'Run' button gets pressed in the environment.
      */
     public void act() 
     {
-        //createDialog();
+        addHiddenSprite();
+        elderDialog();
+        
     }    
     
-    public void createDialog(){
-        if (getOneIntersectingObject(Robot.class) != null){
-            TextPanel helloText = new TextPanel("welcomeMsg");
-            getWorld().addObject(helloText, getWorld().getWidth()/2, getWorld().getHeight()/2 );
-        }
+        protected void addHiddenSprite() {   
+        hs = new HiddenSprite(this, 60, 40, 10, 5, true);  
+        getWorld().addObject(hs, getX(), getY()); 
     }
+    
+     public void elderDialog(){
+        
+        if( hs.getWorld() != null ) {   
+            List<Actor> things = hs.getHitBoxIntersections();    
+            if( things.size() > 1 ) {      
+                int infront = 0;      
+                for(int i=0; i < things.size(); i++ ) {       
+                    Actor a = things.get(i);        
+                    if(a instanceof HiddenSprite)        
+                    continue;        
+                    if( a instanceof Robot) {  
+                         counter--;
+                         if (counter<0 && !isActive){
+                             helloText = new TextPanel("welcomeMsg");
+                             getWorld().addObject(helloText, getWorld().getWidth()/2, getWorld().getHeight()/2);
+                             isActive = true;
+                            }
+                        }
+                        if (Greenfoot.isKeyDown("enter")){
+                            counter = 10;
+                            getWorld().removeObject(helloText);
+                
+                        }
+                        if (a instanceof Elder)
+                        {
+                            //isActive = false;
+            
+                        }       
+                    }
+                }
+            }
+        }
+    
+
 }
