+++ tuto/Robot.ctxt
@@ -6,11 +6,7 @@ comment1.target=void\ act()
 comment1.text=\r\n\ Act\ -\ do\ whatever\ the\ Robot\ wants\ to\ do.\ This\ method\ is\ called\ whenever\r\n\ the\ 'Act'\ or\ 'Run'\ button\ gets\ pressed\ in\ the\ environment.\r\n
 comment2.params=moveAmt
 comment2.target=void\ move(int)
-comment3.params=
-comment3.target=void\ addHiddenSprite()
-comment4.params=
-comment4.target=void\ elderDialog()
-comment5.params=clss
-comment5.target=boolean\ canSee(java.lang.Class)
-comment5.text=\r\n\ Return\ true\ if\ we\ can\ see\ an\ object\ of\ class\ 'clss'\ right\ where\ we\ are.\ \r\n\ False\ if\ there\ is\ no\ such\ object\ here.\r\n
-numComments=6
+comment3.params=clss
+comment3.target=boolean\ canSee(java.lang.Class)
+comment3.text=\r\n\ Return\ true\ if\ we\ can\ see\ an\ object\ of\ class\ 'clss'\ right\ where\ we\ are.\ \r\n\ False\ if\ there\ is\ no\ such\ object\ here.\r\n
+numComments=4
