+++ tuto/TextPanel.ctxt
@@ -8,6 +8,6 @@ comment3.params=
 comment3.target=void\ makeText()
 comment4.params=title
 comment4.target=void\ makeImage(java.lang.String)
-comment5.params=title\ prefix\ iScore
-comment5.target=void\ makeImage(java.lang.String,\ java.lang.String,\ int)
+comment5.params=title\ icon
+comment5.target=void\ makeImage(java.lang.String,\ java.lang.String)
 numComments=6
